+++ R/bed_jaccard.r
@@ -46,6 +46,10 @@
 #'
 #' @export
 bed_jaccard <- function(x, y) {
+
+  check_required(x)
+  check_required(y)
+
   x <- check_interval(x)
   y <- check_interval(y)