+++ R/read_genome.r
@@ -23,8 +23,9 @@
 #'
 #' @export
 read_genome <- function(path) {
+  check_required(path)
   colnames <- c("chrom", "size")
-  genome <- suppressMessages(readr::read_tsv(path, col_names = colnames))
+  genome <- readr::read_tsv(path, col_names = colnames, show_col_types = FALSE)
   genome <- arrange(genome, desc(size))
   genome
 }
@@ -84,10 +85,10 @@ bound_intervals <- function(x, genome, trim = FALSE) {
   }
 
   if (any(res$start == res$end)) {
-    warning(paste0(
-      sum(res$start == res$end),
-      " intervals generated with same start and end were discarded"
-    ))
+    n <- sum(res$start == res$end)
+    cli::cli_warn(
+      "{n} interval{?s} discarded with same start and end after bounding"
+    )
   }
 
   res <- res[res$start != res$end, ]