+++ tests/testthat/test-db.r
@@ -1,5 +1,3 @@
-context("db")
-
 has_internet <- function() {
   !is.null(curl::nslookup("r-project.org", error = FALSE))
 }