+++ tests/testthat/test_genome.r
@@ -1,12 +1,10 @@
-context("genome files")
-
 genome_path <- system.file("extdata", "hg19.chrom.sizes.gz", package = "valr")
 genome <- read_genome(genome_path)
 
 test_that("genomes are correctly read", {
   expect_equal(nrow(genome), 25)
   expect_equal(colnames(genome), c("chrom", "size"))
-  expect_is(genome, "data.frame")
+  expect_true("data.frame" %in% class(genome))
 })
 
 test_that("unbounded intervals are removed", {