+++ tests/testthat/test_makewindows.r
@@ -1,5 +1,3 @@
-context("bed_makewindows")
-
 x <- tibble::tribble(
   ~chrom, ~start, ~end, ~name,
   "chr1", 100, 200, "A",
@@ -84,12 +82,9 @@ test_that("num_win rev", {
 
 test_that("interval is smaller than n windows", {
   # test warning
-  expect_warning(
-    bed_makewindows(x, num_win = 150),
-    "Interval [^:]+:\\d+-\\d+, smaller than requested number of windows. skipping"
-  )
+  expect_message(bed_makewindows(x, num_win = 150))
   # test that intervals are dropped if num_win > than interval size
-  res <- suppressWarnings(bed_makewindows(x, num_win = 150))
+  res <- suppressMessages(bed_makewindows(x, num_win = 150))
   expect_equal(nrow(res), 0)
 })